Changed wrong generated password
Password generated with command secret create contained unscaped $$
character, so shell substituted with the process PID at the creation.

Substituted with another password without $ characters, but, in case
special characters are used, should be scaped with / or ' '

To fix the wrong password, persistent volume had to be deleted, because
password was stored at the database.